@@ -36,8 +36,9 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_9
-        targetCompatibility JavaVersion.VERSION_1_9
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+
     }
 }
 
@@ -61,6 +62,9 @@ dependencies {
     implementation 'com.google.firebase:firebase-crashlytics:18.3.6'
     implementation 'com.google.firebase:firebase-analytics:21.2.1'
     implementation 'com.google.firebase:firebase-perf:20.3.1'
+    implementation 'com.google.firebase:firebase-storage:20.1.0'
+    implementation 'com.google.firebase:firebase-firestore:24.4.5'
+    implementation 'com.google.firebase:firebase-messaging:23.1.2'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
@@ -82,12 +86,14 @@ dependencies {
     /*implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'*/
 
     // Android Room Database
-    annotationProcessor 'androidx.room:room-compiler:2.5.0'
-    implementation 'androidx.room:room-common:2.5.0'
-    implementation 'androidx.room:room-runtime:2.5.0'
-    implementation 'androidx.room:room-testing:2.5.0'
+    annotationProcessor 'androidx.room:room-compiler:2.5.1'
+    implementation 'androidx.room:room-common:2.5.1'
+    implementation 'androidx.room:room-runtime:2.5.1'
+    implementation 'androidx.room:room-testing:2.5.1'
 
     implementation "androidx.cardview:cardview:1.0.0"
 
-    implementation 'com.github.antvis:F2Native:latest.release'
+    implementation 'com.github.antvis:F2Native:2.1.16'
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
+
 }
