@@ -4,6 +4,7 @@ import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 import androidx.room.Transaction;
 import androidx.room.Update;
@@ -18,7 +19,7 @@ import ca.lambton.habittracker.habit.model.Progress;
 @Dao
 public abstract class HabitDao {
 
-    @Insert
+    @Insert(onConflict = OnConflictStrategy.IGNORE)
     public abstract long insertHabit(Habit habit);
 
     @Insert
@@ -35,6 +36,18 @@ public abstract class HabitDao {
 
     @Query("SELECT * FROM HABIT_TBL")
     public abstract LiveData<List<Habit>> getAllHabits();
+
+    @Query("SELECT * FROM HABIT_TBL WHERE HABIT_TYPE = 'PERSONAL' and USER_ID != -1")
+    public abstract LiveData<List<Habit>> getAllPersonalHabits();
+
+    @Query("SELECT * FROM HABIT_TBL WHERE HABIT_TYPE = 'PERSONAL' AND USER_ID = :userId")
+    public abstract LiveData<List<Habit>> getAllPersonalHabitsByUserId(String userId);
+
+    @Query("SELECT * FROM HABIT_TBL WHERE HABIT_TYPE = 'PUBLIC'")
+    public abstract LiveData<List<Habit>> getAllPublicHabits();
+
+    @Query("SELECT * FROM HABIT_TBL WHERE HABIT_TYPE = 'PUBLIC' AND USER_ID = :userId")
+    public abstract LiveData<List<Habit>> getAllPublicHabitsByUserId(String userId);
 
     @Query("SELECT * FROM HABIT_TBL WHERE HABIT_ID = :id")
     public abstract LiveData<Habit> fetchById(Long id);
