@@ -22,7 +22,7 @@ public class Habit implements Serializable {
     @ColumnInfo(name = "PREDEFINED")
     private boolean isPredefined;
     @ColumnInfo(name = "USER_ID")
-    private long userId;
+    private String userId;
     @ColumnInfo(name = "DURATION")
     private int duration;
     @ColumnInfo(name = "DURATION_UNIT")
@@ -90,11 +90,12 @@ public class Habit implements Serializable {
         isPredefined = predefined;
     }
 
-    public long getUserId() {
+
+    public String getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(String userId) {
         this.userId = userId;
     }
 
@@ -168,5 +169,26 @@ public class Habit implements Serializable {
 
     public void setEndDate(long endDate) {
         this.endDate = endDate;
+    }
+
+    @Override
+    public String toString() {
+        return "Habit{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", description='" + description + '\'' +
+                ", creationDate=" + creationDate +
+                ", isPredefined=" + isPredefined +
+                ", userId='" + userId + '\'' +
+                ", duration=" + duration +
+                ", durationUnit='" + durationUnit + '\'' +
+                ", frequency=" + frequency +
+                ", frequencyUnit='" + frequencyUnit + '\'' +
+                ", categoryId=" + categoryId +
+                ", imagePath='" + imagePath + '\'' +
+                ", habitType='" + habitType + '\'' +
+                ", startDate=" + startDate +
+                ", endDate=" + endDate +
+                '}';
     }
 }
