@@ -1,8 +1,13 @@
 package ca.lambton.habittracker.util;
 
 import android.content.Context;
+import android.content.res.AssetManager;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.net.Uri;
 
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.text.SimpleDateFormat;
@@ -10,20 +15,21 @@ import java.util.Date;
 import java.util.Locale;
 
 public class Utils {
-    public static String loadAssetFile(Context context, String assetFile) {
+    public static byte[] loadAssetFile(Context context, String assetFile) {
         try {
             InputStream is = context.getAssets().open(assetFile);
 
             byte[] buf = new byte[1024 * 500];
 
             ByteArrayOutputStream output = new ByteArrayOutputStream();
-            int size = 0;
+            int size;
             while ((size = is.read(buf, 0, buf.length)) >= 0) {
                 output.write(buf, 0, size);
             }
 
-            return new String(output.toByteArray());
+            return output.toByteArray();
         } catch (IOException e) {
+            System.err.println(e.getMessage());
         }
         return null;
     }
@@ -33,4 +39,24 @@ public class Utils {
         Date date = new Date(time);
         return formatter.format(date);
     }
+
+    public static String capitalize(String string) {
+        String firstCharacter = string.substring(0, 1);
+        return firstCharacter.toUpperCase() + string.substring(1).toLowerCase();
+    }
+
+    public static byte[] convertImageToBytes(Context context, String imagePath) throws IOException {
+        InputStream is = context.getAssets().open("img/" + imagePath);
+
+        Bitmap bitmap = BitmapFactory.decodeStream(is);
+        ByteArrayOutputStream stream = new ByteArrayOutputStream();
+        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
+        return stream.toByteArray();
+    }
+
+    public static Bitmap convertBytesToImage(byte[] imageData) {
+        return BitmapFactory.decodeByteArray(imageData, 0, imageData.length);
+    }
+
+
 }
