@@ -10,10 +10,14 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 import ca.lambton.habittracker.R;
 import ca.lambton.habittracker.databinding.FragmentTodayProgressBinding;
@@ -27,10 +31,16 @@ public class SummarizedProgressFragment extends Fragment {
     private FragmentTodayProgressBinding binding;
     private HabitViewModel habitViewModel;
     private float finalResult;
-
     private float todayProgress = 0;
     private float totalFrequencies;
     private final List<HabitProgress> habitProgresses = new ArrayList<>();
+    private FirebaseUser mUser;
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        mUser = FirebaseAuth.getInstance().getCurrentUser();
+    }
 
     @Nullable
     @Override
@@ -49,14 +59,17 @@ public class SummarizedProgressFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+        LocalDate todayDate = LocalDate.now();
 
         habitViewModel.getAllProgress().observe(getViewLifecycleOwner(), habitProgresses1 -> {
-            //System.out.println("My Progress: " + habitProgresses1.size());
-            LocalDate todayDate = LocalDate.now();
             AtomicInteger index = new AtomicInteger();
             habitProgresses.clear();
 
-            for (HabitProgress hp : habitProgresses1) {
+            List<HabitProgress> myHabitProgressFiltered = habitProgresses1.stream()
+                    .filter(dbUser -> dbUser.getHabit().getUserId().equals(mUser.getUid()))
+                    .collect(Collectors.toList());
+
+            for (HabitProgress hp : myHabitProgressFiltered) {
 
                 String startDateString = Utils.parseDate(hp.getHabit().getStartDate());
                 String endDateString = Utils.parseDate(hp.getHabit().getEndDate());
@@ -82,11 +95,11 @@ public class SummarizedProgressFragment extends Fragment {
                 LocalDate endDate = LocalDate.parse(endDateString);
 
                 if (todayDate.isEqual(startDate) || todayDate.isAfter(startDate) && (todayDate.isEqual(endDate) || (todayDate.isBefore(endDate)))) {
-                    if (habitProgresses1.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.DAILY.name())) {
+                    if (myHabitProgressFiltered.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.DAILY.name())) {
                         todayProgress += habitProgress.getProgressList().stream().filter(progress -> progress.getDate().equals(todayDate.toString())).map(Progress::getCounter).mapToInt(Integer::intValue).sum();
-                    } else if (habitProgresses1.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.WEEKLY.name())) {
+                    } else if (myHabitProgressFiltered.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.WEEKLY.name())) {
                         todayProgress += habitProgress.getProgressList().stream().filter(progress -> progress.getDate().equals(todayDate.toString())).map(Progress::getCounter).mapToInt(Integer::intValue).sum();
-                    } else if (habitProgresses1.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.MONTHLY.name())) {
+                    } else if (myHabitProgressFiltered.get(index.get()).getHabit().getFrequencyUnit().equalsIgnoreCase(Frequency.MONTHLY.name())) {
                         todayProgress += habitProgress.getProgressList().stream().filter(progress -> progress.getDate().equals(todayDate.toString())).map(Progress::getCounter).mapToInt(Integer::intValue).sum();
                     }
                 }
