@@ -15,6 +15,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.antgroup.antv.f2.F2CanvasView;
 import com.antgroup.antv.f2.F2Chart;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -31,7 +33,6 @@ import ca.lambton.habittracker.databinding.FragmentRecycleViewBinding;
 import ca.lambton.habittracker.habit.model.HabitProgress;
 import ca.lambton.habittracker.habit.model.Progress;
 import ca.lambton.habittracker.habit.viewmodel.HabitViewModel;
-import ca.lambton.habittracker.habit.viewmodel.HabitViewModelFactory;
 import ca.lambton.habittracker.util.Utils;
 
 public class TodayReportFragment extends Fragment {
@@ -40,88 +41,18 @@ public class TodayReportFragment extends Fragment {
     private TodayReportAdapter todayReportAdapter;
     private final List<HabitProgress> habitProgressList = new ArrayList<>();
     private HabitViewModel habitViewModel;
+    private FirebaseUser mUser;
+    private RecyclerView recycleView;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         binding = FragmentRecycleViewBinding.inflate(LayoutInflater.from(requireContext()));
-        RecyclerView recycleView = binding.recycleView;
-        habitViewModel = new ViewModelProvider(requireActivity(), new HabitViewModelFactory(requireActivity().getApplication())).get(HabitViewModel.class);
+        recycleView = binding.recycleView;
+        habitViewModel = new ViewModelProvider(requireActivity()).get(HabitViewModel.class);
 
-        todayReportAdapter = new TodayReportAdapter(habitProgressList, (canvasView, position) -> {
-
-            canvasView.setAdapter(new F2CanvasView.Adapter() {
-                F2Chart mChart;
-
-                @Override
-                public void onCanvasDraw(F2CanvasView canvasView) {
-
-                    //Initialize the chart, because you need the width and height of the canvasView,
-                    //initialize here to ensure that you can get the width and height of the canvasView
-
-                    if (mChart == null) {
-                        mChart = F2Chart.create(canvasView.getContext(), "Report", canvasView.getWidth(), canvasView.getHeight());
-                    }
-                    // 700 x 350
-                    //Associating the chart and canvasView, the chart finally needs to display the chart on the canvasView
-                    mChart.setCanvas(canvasView);
-                    //set chart padding
-                    mChart.padding(10.0, 10.0, 10.0, 10.0);
-                    //load data from data.json
-
-                    //TODO: obtain all records in one day
-                    JSONArray jsonArray = new JSONArray();
-
-                    AtomicInteger counter = new AtomicInteger();
-                    AtomicInteger value = new AtomicInteger();
-
-                    Map<String, Integer> groupByDate = habitProgressList.get(position).getProgressList().stream()
-                            .collect(Collectors.groupingBy(Progress::getDate, Collectors.summingInt(Progress::getCounter)));
-
-
-                    groupByDate.forEach((date, totalPerDay) -> {
-                        value.getAndIncrement();
-
-                        JSONObject jsonObject = new JSONObject();
-                        try {
-
-                            jsonObject.put("progress", totalPerDay);
-                            jsonObject.put("date", date);
-
-                            jsonArray.put(jsonObject);
-                        } catch (JSONException e) {
-                            throw new RuntimeException(e);
-                        }
-                    });
-
-                    //set data to chart
-                    String source = jsonArray.toString();
-                    System.out.println(source);
-                    mChart.source(source);
-
-                    mChart.setScale("progress", new F2Chart.ScaleConfigBuilder().precision(0).max(habitProgressList.get(position).getHabit().getFrequency()).min(0));
-                    mChart.setScale("date", new F2Chart.ScaleConfigBuilder().precision(0).tickCount(2).range(new double[]{0.1, 0.9}));
-
-                    //Draw a polyline on the chart. The data mappings of the x-axis and y-axis of the polyline are genre and sold fields respectively.
-                    mChart.line().color("type", new String[]{"ORANGE"})
-                            .fixedSize(2.f)
-                            .fixedShape("smooth")
-                            .position("date*progress");
-
-                    //Render and display on canvasView
-                    mChart.render();
-                }
-
-                @Override
-                public void onDestroy() {
-                    if (mChart != null) {
-                        mChart.destroy();
-                    }
-                }
-            });
-
-        });
+        mUser = FirebaseAuth.getInstance().getCurrentUser();
 
         recycleView.setLayoutManager(new GridLayoutManager(requireContext(), 2, LinearLayoutManager.VERTICAL, false));
         recycleView.setAdapter(todayReportAdapter);
@@ -130,11 +61,88 @@ public class TodayReportFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        todayReportAdapter = new TodayReportAdapter(habitProgressList, (canvasView, position) -> canvasView.setAdapter(new F2CanvasView.Adapter() {
+            F2Chart mChart;
+
+            @Override
+            public void onCanvasDraw(F2CanvasView canvasView) {
+
+                //Initialize the chart, because you need the width and height of the canvasView,
+                //initialize here to ensure that you can get the width and height of the canvasView
+                if (mChart == null) {
+                    mChart = F2Chart.create(canvasView.getContext(), "Report", canvasView.getWidth(), canvasView.getHeight());
+                }
+                // 700 x 350
+                //Associating the chart and canvasView, the chart finally needs to display the chart on the canvasView
+                mChart.setCanvas(canvasView);
+                //set chart padding
+                mChart.padding(10.0, 10.0, 10.0, 10.0);
+                //load data from data.json
+
+                JSONArray jsonArray = new JSONArray();
+
+                AtomicInteger value = new AtomicInteger();
+
+                Map<String, Integer> groupByDate = habitProgressList.get(position).getProgressList().stream()
+                        .collect(Collectors.groupingBy(Progress::getDate, Collectors.summingInt(Progress::getCounter)));
+
+
+                groupByDate.forEach((date, totalPerDay) -> {
+                    value.getAndIncrement();
+
+                    JSONObject jsonObject = new JSONObject();
+                    try {
+
+                        jsonObject.put("progress", totalPerDay);
+                        jsonObject.put("date", date);
+
+                        jsonArray.put(jsonObject);
+                    } catch (JSONException e) {
+                        throw new RuntimeException(e);
+                    }
+                });
+
+                //set data to chart
+                String source = jsonArray.toString();
+                System.out.println(source);
+                mChart.source(source);
+
+                mChart.setScale("progress", new F2Chart.ScaleConfigBuilder().precision(0).max(habitProgressList.get(position).getHabit().getFrequency()).min(0));
+                mChart.setScale("date", new F2Chart.ScaleConfigBuilder().precision(0).tickCount(2).range(new double[]{0.1, 0.9}));
+
+                //Draw a polyline on the chart. The data mappings of the x-axis and y-axis of the polyline are genre and sold fields respectively.
+                mChart.line().color("type", new String[]{"ORANGE"})
+                        .fixedSize(2.f)
+                        .fixedShape("smooth")
+                        .position("date*progress");
+
+                //Render and display on canvasView
+                mChart.render();
+            }
+
+            @Override
+            public void onDestroy() {
+                if (mChart != null) {
+                    mChart.destroy();
+                }
+            }
+        }));
+
+        return binding.getRoot();
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+
         LocalDate todayDate = LocalDate.now();
 
         habitViewModel.getAllProgress().observe(requireActivity(), habitProgresses -> {
+
+            List<HabitProgress> habitProgressFiltered = habitProgresses.stream().filter(dbUser -> dbUser.getHabit().getUserId().equals(mUser.getUid())).collect(Collectors.toList());
+
             habitProgressList.clear();
-            habitProgresses.forEach(habitProgress -> {
+            habitProgressFiltered.forEach(habitProgress -> {
                 String startDateString = Utils.parseDate(habitProgress.getHabit().getStartDate());
                 String endDateString = Utils.parseDate(habitProgress.getHabit().getEndDate());
 
@@ -146,10 +154,9 @@ public class TodayReportFragment extends Fragment {
                 }
             });
 
-            todayReportAdapter.notifyDataSetChanged();
+            todayReportAdapter.notifyItemChanged(0, habitProgressFiltered.size());
         });
+        recycleView.setAdapter(todayReportAdapter);
 
-        return binding.getRoot();
     }
-
 }
