@@ -4,20 +4,29 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.NavDirections;
+import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.android.material.progressindicator.CircularProgressIndicator;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import ca.lambton.habittracker.R;
 import ca.lambton.habittracker.databinding.FragmentGroupHabitsBinding;
 import ca.lambton.habittracker.habit.model.Habit;
+import ca.lambton.habittracker.habit.model.HabitProgress;
 import ca.lambton.habittracker.habit.viewmodel.HabitViewModel;
-import ca.lambton.habittracker.habit.viewmodel.HabitViewModelFactory;
 
 public class GroupHabitsFragment extends Fragment {
 
@@ -26,38 +35,22 @@ public class GroupHabitsFragment extends Fragment {
 
     HabitViewModel habitViewModel;
 
-    private final List<Habit> habits = new ArrayList<>();
+    private FirebaseUser mUser;
 
-    public GroupHabitsFragment() {
-        // Required empty public constructor
-    }
-
-    public static GroupHabitsFragment newInstance(String param1, String param2) {
-        GroupHabitsFragment fragment = new GroupHabitsFragment();
-        Bundle args = new Bundle();
-        fragment.setArguments(args);
-        return fragment;
-    }
+    private final List<HabitProgress> habits = new ArrayList<>();
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-        }
+        binding = FragmentGroupHabitsBinding.inflate(LayoutInflater.from(requireContext()));
+        habitViewModel = new ViewModelProvider(requireActivity()).get(HabitViewModel.class);
+        mUser = FirebaseAuth.getInstance().getCurrentUser();
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        binding = FragmentGroupHabitsBinding.inflate(inflater, container, false);
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         RecyclerView groupRecyclerView = binding.groupOngoingHabitsList;
 
-        habitViewModel = new ViewModelProvider(this, new HabitViewModelFactory(getActivity().getApplication())).get(HabitViewModel.class);
-        habitViewModel.getAllHabit().observe(getViewLifecycleOwner(), result -> {
-            this.habits.clear();
-            this.habits.addAll(result);
-            groupOngoingHabitListAdapter.notifyDataSetChanged();
-        });
 
         groupOngoingHabitListAdapter = new OngoingHabitsRecycleAdapter(habits, getOnCallbackOngoingHabit(habits, true), this.getContext(), true);
         groupRecyclerView.setLayoutManager(new GridLayoutManager(getContext(), 2));
@@ -67,13 +60,38 @@ public class GroupHabitsFragment extends Fragment {
     }
 
     @NonNull
-    private OngoingHabitsRecycleAdapter.OnOngoingHabitsCallback getOnCallbackOngoingHabit(List<Habit> habits, Boolean isGroup) {
+    private OngoingHabitsRecycleAdapter.OnOngoingHabitsCallback getOnCallbackOngoingHabit(List<HabitProgress> habits, boolean isGroup) {
         return new OngoingHabitsRecycleAdapter.OnOngoingHabitsCallback() {
-
             @Override
             public void onRowClicked(int position, boolean isGroup) {
+                if (isGroup) {
+                    Navigation.findNavController(getView()).navigate(R.id.groupHabitDetailFragment);
+                } else {
+                    NavDirections navDirections = AllHabitsFragmentDirections.actionAllHabitsFragmentToPrivateHabitDetailFragment(null).setHabitProgress(habits.get(position));
+                    Navigation.findNavController(requireActivity(), R.id.nav_host_fragment_content_main).navigate(navDirections);
+                }
+            }
+
+            @Override
+            public void getProgressList(TextView habitPercentageNumText, CircularProgressIndicator habitProgressbar, int totalTimesToComplete, int position) {
 
             }
         };
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        this.habits.clear();
+
+        habitViewModel.getAllProgress().observe(this, habitProgressResult -> {
+
+            List<HabitProgress> resultData = habitProgressResult.stream()
+                    .filter(progress -> progress.getHabit().getUserId().equals(mUser.getUid()))
+                    .collect(Collectors.toList());
+
+            habits.addAll(resultData);
+            groupOngoingHabitListAdapter.notifyDataSetChanged();
+        });
     }
 }
