@@ -10,25 +10,32 @@ import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.android.material.progressindicator.CircularProgressIndicator;
+
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 import ca.lambton.habittracker.R;
 import ca.lambton.habittracker.habit.model.Habit;
+import ca.lambton.habittracker.habit.model.HabitProgress;
 
 public class OngoingHabitsRecycleAdapter extends RecyclerView.Adapter<OngoingHabitsRecycleAdapter.ViewHolder> {
 
     private final OnOngoingHabitsCallback onOngoingHabitsCallback;
-    private final List<Habit> habits;
+    private final List<HabitProgress> habits;
     private Context context;
 
     private Boolean isGroup;
 
-    public OngoingHabitsRecycleAdapter(List<Habit> habits, OnOngoingHabitsCallback onCallback) {
+    public OngoingHabitsRecycleAdapter(List<HabitProgress> habits, OnOngoingHabitsCallback onCallback) {
         this.habits = habits;
         this.onOngoingHabitsCallback = onCallback;
     }
 
-    public OngoingHabitsRecycleAdapter(List<Habit> habits, OnOngoingHabitsCallback onCallback, Context context, Boolean isGroup) {
+    public OngoingHabitsRecycleAdapter(List<HabitProgress> habits, OnOngoingHabitsCallback onCallback, Context context, Boolean isGroup) {
         this.habits = habits;
         this.onOngoingHabitsCallback = onCallback;
         this.context = context;
@@ -47,8 +54,33 @@ public class OngoingHabitsRecycleAdapter extends RecyclerView.Adapter<OngoingHab
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
 
-        holder.habitNameLabel.setText(habits.get(position).getName());
-        holder.habitPercentageNumText.setText("73%");
+        int totalTimesToComplete = 0;
+        String frequencyUnit = habits.get(position).getHabit().getFrequencyUnit();
+        int frequencyValue = habits.get(position).getHabit().getFrequency();
+        long startDateMillis = habits.get(position).getHabit().getStartDate();
+        long endDateMillis = habits.get(position).getHabit().getEndDate();
+
+        LocalDate startDate = Instant.ofEpochMilli(startDateMillis).atZone(ZoneId.systemDefault()).toLocalDate();
+        LocalDate endDate = Instant.ofEpochMilli(endDateMillis).atZone(ZoneId.systemDefault()).toLocalDate();
+        long daysBetween = ChronoUnit.DAYS.between(startDate, endDate);
+
+        if (frequencyUnit.equals("DAILY")) {
+            totalTimesToComplete = frequencyValue * (int) daysBetween;
+            onOngoingHabitsCallback.getProgressList(holder.habitPercentageNumText, holder.habitProgressbar, totalTimesToComplete, position);
+        } else if (frequencyUnit.equals("WEEKLY")) {
+            totalTimesToComplete = frequencyValue * ((int) daysBetween / 7);
+            onOngoingHabitsCallback.getProgressList(holder.habitPercentageNumText, holder.habitProgressbar, totalTimesToComplete, position);
+        } else {
+            if (daysBetween == 30) {
+                totalTimesToComplete = frequencyValue * 30;
+            }
+            else {
+                totalTimesToComplete = frequencyValue * ((int) daysBetween / 30);
+            }
+            onOngoingHabitsCallback.getProgressList(holder.habitPercentageNumText, holder.habitProgressbar, totalTimesToComplete, position);
+        }
+
+        holder.habitNameLabel.setText(habits.get(position).getHabit().getName());
         holder.ongoingHabitCard.setOnClickListener(view -> {
             onOngoingHabitsCallback.onRowClicked(position, isGroup);
         });
@@ -72,6 +104,8 @@ public class OngoingHabitsRecycleAdapter extends RecyclerView.Adapter<OngoingHab
 
         private final TextView habitPercentageNumText;
 
+        private final CircularProgressIndicator habitProgressbar;
+
         private final CardView ongoingHabitCard;
 
         public ViewHolder(@NonNull View itemView) {
@@ -81,10 +115,13 @@ public class OngoingHabitsRecycleAdapter extends RecyclerView.Adapter<OngoingHab
             memberCountLabel = itemView.findViewById(R.id.memberCountLabel);
             habitPercentageNumText = itemView.findViewById(R.id.habitPercentageNumText);
             ongoingHabitCard = itemView.findViewById(R.id.ongoingHabitCard);
+            habitProgressbar = itemView.findViewById(R.id.habitProgressbar);
         }
     }
 
     public interface OnOngoingHabitsCallback {
         void onRowClicked(int position, boolean isGroup);
+
+        void getProgressList(TextView habitPercentageNumText, CircularProgressIndicator habitProgressbar, int totalTimesToComplete, int position);
     }
 }
